bot/requirements.txt
python-telegram-bot==13.11
loguru
python-dotenv
sqlalchemy
psycopg2
sqlalchemy-utils
requests